+++ record/android/src/main/java/com/llfbandit/record/RecordPlugin.java
@@ -6,14 +6,20 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodChannel;
 
 /**
  * RecordPlugin
  */
 public class RecordPlugin implements FlutterPlugin, ActivityAware {
+  private static final String MESSAGES_CHANNEL = "com.llfbandit.record/messages";
+  private static final String EVENTS_CHANNEL = "com.llfbandit.record/events";
+
   /// The MethodChannel that will the communication between Flutter and native Android
-  private MethodChannel channel;
+  private MethodChannel methodChannel;
+  // Channel for communicating with flutter using async stream
+  private EventChannel eventChannel;
   /// Our call handler
   private MethodCallHandlerImpl handler;
   private FlutterPluginBinding pluginBinding;
@@ -63,19 +69,24 @@ public class RecordPlugin implements FlutterPlugin, ActivityAware {
   private void startPlugin(BinaryMessenger messenger, ActivityPluginBinding binding) {
 
     handler = new MethodCallHandlerImpl(binding.getActivity());
-    channel = new MethodChannel(messenger, "com.llfbandit.record");
-    channel.setMethodCallHandler(handler);
+    methodChannel = new MethodChannel(messenger, MESSAGES_CHANNEL);
+    methodChannel.setMethodCallHandler(handler);
 
     binding.addRequestPermissionsResultListener(handler);
+
+    eventChannel = new EventChannel(messenger, EVENTS_CHANNEL);
+    eventChannel.setStreamHandler(handler);
   }
 
   private void stopPlugin() {
     activityBinding.removeRequestPermissionsResultListener(handler);
     activityBinding = null;
-    channel.setMethodCallHandler(null);
+    methodChannel.setMethodCallHandler(null);
+    eventChannel.setStreamHandler(null);
     handler.close();
     handler = null;
-    channel = null;
+    methodChannel = null;
+    eventChannel = null;
   }
   /// END ActivityAware
   /////////////////////////////////////////////////////////////////////////////